+++ lib/src/about_content.dart
@@ -124,26 +124,10 @@ class _AboutContentState extends State<AboutContent> {
 
   @override
   Widget build(BuildContext context) {
-    TextStyle header;
-    TextStyle bodyText;
-    TextStyle versionStyle;
-    TextStyle copyrightStyle;
-
-    if (isCupertino(context)) {
-      final theme = CupertinoTheme.of(context).textTheme;
-      header = theme.navLargeTitleTextStyle;
-      versionStyle = theme.textStyle;
-      bodyText = theme.textStyle;
-      copyrightStyle = theme.textStyle;
-    } else {
-      final theme = Theme.of(context).textTheme;
-      header = theme.headline5;
-      bodyText = theme.bodyText2;
-      copyrightStyle = theme.caption;
-    }
-
     final name = widget.applicationName ?? defaultApplicationName(context);
+
     final icon = widget.applicationIcon ?? defaultApplicationIcon(context);
+
     final body = <Widget>[];
 
     if (icon != null) {
@@ -165,12 +149,12 @@ class _AboutContentState extends State<AboutContent> {
             children: <Widget>[
               Text(
                 name,
-                style: header,
+                style: Theme.of(context).textTheme.headline5,
                 textAlign: TextAlign.center,
               ),
               Text(
                 version,
-                style: versionStyle,
+                style: Theme.of(context).textTheme.bodyText2,
                 textAlign: TextAlign.center,
               ),
               if (widget.applicationLegalese != null)
@@ -178,15 +162,25 @@ class _AboutContentState extends State<AboutContent> {
                   padding: const EdgeInsets.only(top: 18),
                   child: Text(
                     Template(widget.applicationLegalese).render(_values),
-                    style: copyrightStyle,
+                    style: Theme.of(context).textTheme.caption,
                     textAlign: TextAlign.center,
                   ),
                 ),
               if (widget.applicationDescription != null) Divider(),
               if (widget.applicationDescription != null)
+                // Padding(
+                // padding: const EdgeInsets.only(top: 18),
+                // child:
                 Container(
                     padding: const EdgeInsets.only(top: 8, bottom: 8),
+                    // decoration: const BoxDecoration(
+                    //   border: Border(
+                    //     top: BorderSide(width: 0.4),
+                    //     bottom: BorderSide(width: 0.4),
+                    //   ),
+                    // ),
                     child: widget.applicationDescription),
+              // ),
               if (widget.applicationDescription != null) Divider(),
             ],
           ),
@@ -200,10 +194,7 @@ class _AboutContentState extends State<AboutContent> {
 
     return Scrollbar(
       child: SingleChildScrollView(
-        child: DefaultTextStyle(
-          style: bodyText,
-          child: ListBody(children: body),
-        ),
+        child: ListBody(children: body),
       ),
     );
   }